split Present/FullCard to Cart/PresentCar
